Rename /models to /entities
