+++ data/lib/entities/app_open_meta.dart
