+++ data/lib/entities/app_update.dart
