+++ data/lib/entities/app_update_state.dart
