+++ data/lib/entities/localization_data.dart
