+++ data/lib/entities/message_localization.dart
