+++ data/lib/entities/nstack_app_open_data.dart
