+++ data/lib/entities/platform.dart
