+++ data/lib/local_repository.dart
@@ -1,4 +1,4 @@
-import 'package:data/models/localize_resource_list.dart';
+import 'package:data/entities/localize_resource_list.dart';
 
 abstract class LocalRepository {
   Future<void> setLocalization(
