+++ example/lib/nstack.dart
@@ -1,8 +1,8 @@
 /// Generated by NStack, do not modify this file.
 
 import 'package:flutter/widgets.dart';
-import 'package:data/models/language.dart';
-import 'package:data/models/nstack_config.dart';
+import 'package:data/entities/language.dart';
+import 'package:data/entities/nstack_config.dart';
 import 'package:nstack/nstack.dart';
 import 'package:nstack/section_key_delegate.dart';
