+++ lib/src/extensions/localization_resource_extension.dart
@@ -1,4 +1,4 @@
-import 'package:data/models/localize_resource.dart';
+import 'package:data/entities/localize_resource.dart';
 
 extension LocalizationResourceExtension on LocalizeResource {
   String getLocalizedSectionKey(String section, String key) {
