+++ lib/src/nstack_builder.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:build/build.dart';
-import 'package:data/models/language.dart';
-import 'package:data/models/localization_data.dart';
-import 'package:data/models/localize_index.dart';
-import 'package:data/models/nstack_config.dart';
+import 'package:data/entities/language.dart';
+import 'package:data/entities/localization_data.dart';
+import 'package:data/entities/localize_index.dart';
+import 'package:data/entities/nstack_config.dart';
 import 'package:nstack/src/nstack_repository.dart';
 import 'package:nstack/src/other/extensions.dart';
 import 'package:nstack/src/other/reserved_keywords.dart';
@@ -83,8 +83,8 @@ class NstackBuilder implements Builder {
 /// Generated by NStack, do not modify this file.
 
 import 'package:flutter/widgets.dart';
-import 'package:data/models/language.dart';
-import 'package:data/models/nstack_config.dart';
+import 'package:data/entities/language.dart';
+import 'package:data/entities/nstack_config.dart';
 import 'package:nstack/nstack.dart';
 import 'package:nstack/section_key_delegate.dart';
